refactor: move modal callbacks from PipelinesApp into ModalManager
- PipelinesApp (342 → 259 lines): removed 5 callback handlers, handler query,
  handler details query, updateHandlerMutation, and 11 props passed to ModalManager
- ModalManager (57 → 133 lines): now self-contained with all modal navigation logic:
  - handleSuccess (close modal)
  - handleHandlerSelected (persist handler + open settings)
  - handleChangeHandler (open handler selection)
  - handleOAuthConnect (open OAuth modal)
  - handleBackToSettings (return to handler settings)
- ModalManager fetches its own data (handlers, handlerDetails) via TanStack Query
- PipelinesApp no longer imports useHandlers, useHandlerDetails, useUpdateFlowHandler,
  closeModal, activeModal, or modalData from the store
- ModalManager renders as <ModalManager /> with zero props

The modal system was already using Zustand for state — the callbacks were
unnecessary prop drilling through 3 layers (PipelinesApp → ModalManager → ModalSwitch).

Resolves P1 modal prop drilling item from React frontend tech debt audit.
